% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_academic_year_format.R
\name{fn_academic_year_format}
\alias{fn_academic_year_format}
\title{Format the academic year. This function prevents the need to save the academic year in many different formats under different variables. Instead this function formats the short form of the academic year (removing any dashes) and outputs the requested type of long form academic year}
\usage{
fn_academic_year_format(academic_year_short = NULL, type = NULL)
}
\arguments{
\item{academic_year_short}{Insert the shortened version of the current academic year which needs formatting (eg for the academic year 2021 through 2022 this would be of the form 2122 or 21-22)}

\item{type}{Select the type of formatted output. Use one of: 'Long Dash', 'Long Slash', 'Short Dash' or 'Short Slash' only. (eg for the academic year 2021 through 2022 the following outputs can be achieved: 'Long Dash': 2021-2022, 'Long Slash': 2021/2022, 'Short Dash': 2021-22 or 'Short Slash': 2021/22.)}
}
\value{
Returns the formatted academic year
}
\description{
Format the academic year. This function prevents the need to save the academic year in many different formats under different variables. Instead this function formats the short form of the academic year (removing any dashes) and outputs the requested type of long form academic year
}
\examples{
fn_academic_year_format(academic_year_short = "2122", type= "Long Slash"), fn_academic_year_format("21-22", "Short Dash")

}
