% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_anova.R
\name{fn_anova}
\alias{fn_anova}
\title{Perform ANOVA, save ANOVA results, adjusted means and observed means for different stages}
\usage{
fn_anova(
  data_demog,
  col_score,
  vars_all,
  vars_anova,
  stages = NULL,
  stages_for_filtering = NULL,
  append_name = NULL,
  report_observed_mean_for_all_vars
)
}
\arguments{
\item{data_demog}{A dataframe - containing demographics data columns (independent variables) and the student score (dependent variable) in another column and the relevant stage of each student row (if multiple stages are to be considered)}

\item{col_score}{A string - the name of the column containing the student scores (the dependent variable)}

\item{vars_all}{A vector of strings - all the column names of the demographic properties to consider}

\item{vars_anova}{A vector of strings - all the column names of the demographic properties to consider for the Anova}

\item{stages}{A vector (OPTIONAL, if append_name defined. Must define is Stage is to be one of the ANOVA vars) - denoting the stages to be considered. (stages = 2 or stages = c(2,3,4,5) etc.) If multiple stages are to be considered there must be a 'Stage' col in data_demog. If not defined, no stage separation occurs - instead the user must set the append_name variable to define what string is added to results names. Additionally, if no stage separation but Stage is to be one of the ANOVA variables, 'stages_for_filtering' must be defined.}

\item{stages_for_filtering}{A vector (OPTIONAL, must define if Stage is to be one of the ANOVA vars and stage separation is not requested) - denoting the stages to be considered in the ANOVA. (stages_for_filtering = 2 or stages_for_filtering = c(2,3,4,5) etc.) If multiple stages are to be considered there must be a 'Stage' col in data_demog.}

\item{append_name}{A string (OPTIONAL, REQUIRED if stages is not defined) - this string is appended to all results (only if the stages variable is not set)}

\item{report_observed_mean_for_all_vars}{Boolean (TRUE/FALSE) - TRUE: Report observed means for all the demographics (vars_all). FALSE: Report observed means for only the demographics which were not included in the Anova}
}
\value{
A list of dataframes with the anova results, summarised anova results, adjusted means table and observed means table. Use the append function to add results to tabDemog if multiple fn_anova functions are to be run (see examples).
}
\description{
Perform ANOVA, save ANOVA results, adjusted means and observed means for different stages
}
\examples{
#With stage separation:
    tabDemog <- append(tabDemog,fn_anova(data_demog = dfDemog, stages = cnst$stages, col_score = glue("{cnst$assessment}_Score"), vars_all = c("Gender", "Ethnicity", "Disability", "Entry", "Origin"), vars_anova = c("Gender", "Ethnicity", "Disability"), report_observed_mean_for_all_vars = FALSE))
# Without stage separation:
    tabDemog <- append(tabDemog,fn_anova(data_demog = dfDemog, col_score = glue("{cnst$assessment}_Score"), append_name = "All", vars_all = c("Gender", "Ethnicity", "Disability", "Stage", "Entry", "Origin"), vars_anova = c("Gender", "Ethnicity", "Disability", "Stage"), report_observed_mean_for_all_vars = FALSE)))


}
