% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_check_variables_exist.R
\name{fn_check_variables_exist}
\alias{fn_check_variables_exist}
\title{Check If Variables Exist}
\usage{
fn_check_variables_exist(var_names, message_text = "Please set the variable: ")
}
\arguments{
\item{var_names}{A character vector containing the names of the variables to check.}

\item{message_text}{A character string to customize the alert message. Defaults to "Please set the variable".}
}
\value{
This function does not return any value but prints messages to alert the user to set missing variables.
}
\description{
This function checks whether specified variables exist in the global environment.
If a variable does not exist or is empty, it alerts the user to set the variable.
}
\details{
The function checks if the variables exist in the global environment. For list variables,
it checks if the specified element of the list exists and is non-NULL.
For non-list variables, it checks if they exist and are non-NULL.
}
\examples{
files <- list(data_source = 1)
some_var = 2
fn_check_variables_exist(c("files$data_source", "files$other_var", "some_var"))
# Customizing the message text:
fn_check_variables_exist(c("files$data_source", "files$other_var"), message_text = "Set the following variable:")

}
\seealso{
[exists()], [get()]
}
