% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_document.R
\name{fn_document}
\alias{fn_document}
\title{Run the document command for the PsychometricsGP package}
\usage{
fn_document()
}
\value{
Nothing is returned, a message alerts the user when documenting is complete
}
\description{
Use this function to run the document command when updates are made in the PyschometricsGP package. The required packages (devtools, roxygen2) are both loaded and then the document() command is run.
}
\examples{
fn_document()


}
\author{
Gergo Pinter, \email{gergo.pinter@plymouth.ac.uk}, \url{https://www.gergo.io}
}
