% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_item_review.R
\name{fn_item_review}
\alias{fn_item_review}
\title{Detect Items to Review}
\usage{
fn_item_review(
  assessment_type = NULL,
  item_analysis_data = NULL,
  test_in_year = NULL
)
}
\arguments{
\item{assessment_type}{A string. The type of assessment being analysed. This will determine the selection process for finding and returning items to review}

\item{item_analysis_data}{A dataframe. This dataframe must contain at least the following columns: "Item" (the items number), "Stage 5 Facility" (for ADK), "Stage 5 vs 2 Growth" (for ADK), "Stage 5 PtBis" (for ADK) OR "Stage 3 Facility" (for ADTK), "Stage 3 vs 2 Growth" (for ADTK), "Stage 3 PtBis" (for ADTK) OR "Stage 2 Facility" (for PAPT), "Stage 2 vs 1 Growth" (for PAPT), "Stage 2 PtBis" (for PAPT) OR "Facility" (for IDS/Y1KT), "PtBis" (for IDS/Y1KY)}

\item{test_in_year}{(OPTIONAL unless AMK) - Required for AMK assessments so the stage to use for item review can be determined}
}
\value{
A list of the detected items to review is returned.
}
\description{
Detect Items to Review
}
\examples{
#For ADK/ADTK/PAPT/IDS/Y1KT/AMK/AKT:
 cnst <- append(
 cnst,
  fn_item_review(assessment_type = cnst$assessment_type, item_analysis_data = tab$itemAnalysis)
  )

# For AMK:
 cnst <- append(
 cnst,
  fn_item_review(assessment_type = cnst$assessment_type, item_analysis_data = tab$itemAnalysis, test_in_year = cnst$test_in_year)
  )


}
