% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plt_borderline_regression.R
\name{fn_plt_borderline_regression}
\alias{fn_plt_borderline_regression}
\title{Plot of the Borderline Regression method}
\usage{
fn_plt_borderline_regression(x, grade_col = "Grade", score_col = "Score")
}
\arguments{
\item{x}{A data frame containing the global grade and score data.}

\item{grade_col}{Name of the column containing the global grades.}

\item{score_col}{Name of the column containing the assessment scores.}
}
\value{
A list containing:
\item{plt_borderline_regression}{A ggplot object}
\item{pass_mark}{The calculated pass mark}
Additionally the linear regression model (b_reg_line) is passed to the
global environment.
}
\description{
fn_plt_borderline_regression creates a scatterplot of assessment
scores against global grades. A regression line is added and its intersection
with the borderline group is used to calculate a pass mark.
}
