% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reliability.R
\name{fn_reliability}
\alias{fn_reliability}
\title{Test Reliability Parameters (Non/Stage Specific)}
\usage{
fn_reliability(
  stages = NULL,
  data_scores = NULL,
  data_results = NULL,
  items_included = NULL
)
}
\arguments{
\item{stages}{(Optional) Vector containing all the stages being considered. If undefined, only a single dataset (data_scores) is expected/considered.}

\item{data_scores}{If stage specific: A list of dataframes with students scores for each stage. Each data frame in the list should be named "stage#" where # is the relevant stage.
If non stage specific: A single dataframe containing what would otherwise be contained in one of the dataframes in the list of stage specific dataframes. In any dataframe the columns should be the items and the rows should be the index of each student (actual value unimportant). Each data point should represent the score for a particular student on a particular item. All students with NA scores and excluded items should be removed prior to use in this function}

\item{data_results}{A dataframe containing student results. This variable must contain a "Stage" column in which the stage of each student is entered in a new row. This is used to add the number of students assessed in each Stage to the resultant export of reliability data.}

\item{items_included}{(Integer) The number of items included in the exam. Usually stored under l_exam_vars$n_items_incl}
}
\value{
A dataframe containing the reliability data for each considered stage is returned
}
\description{
Test Reliability Parameters (Non/Stage Specific)
}
\examples{
fn_reliability(stages = cnst$stages, data_scores = df_res_stages_only_scores, items_included = l_exam_vars$n_items_incl)
}
