% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reload.R
\name{fn_reload}
\alias{fn_reload}
\title{Reload the PsychometricsGP package}
\usage{
fn_reload(repo = "GergoIO/psychometricsGP")
}
\arguments{
\item{repo}{The GitHub repository from which to install the package. Defaults to "GergoIO/psychometricsGP".}
}
\value{
Nothing is returned, a message alerts the user when reloading is complete.
}
\description{
Use this function to reload the current package (psychometricsGP). This function is useful when changes have been made and uploaded to the package repository. This function takes no arguments.
}
\examples{
fn_reload()
}
