% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_report_add_paragraphs.R
\name{fn_report_add_paragraphs}
\alias{fn_report_add_paragraphs}
\title{Add Multiple Paragraphs to a Word Document}
\usage{
fn_report_add_paragraphs(report, content_list)
}
\arguments{
\item{report}{An `officer::rdocx` object representing the Word document.}

\item{content_list}{A list or vector of character strings, where each element
represents a paragraph to be added to the document.}
}
\value{
The modified `rdocx` object with the added paragraphs.
}
\description{
This function adds multiple paragraphs to a Word document report object.
}
\examples{
\dontrun{
library(officer)
doc <- read_docx()
content <- c("This is the first paragraph.", "This is the second paragraph.")
doc <- fn_report_add_paragraphs(doc, content)
print(doc, target = "output.docx")
}
}
