% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_response_summary.R
\name{fn_response_summary}
\alias{fn_response_summary}
\title{Response Summary Table}
\usage{
fn_response_summary(stages = NULL, results = NULL)
}
\arguments{
\item{stages}{A vector - denoting the stages to be considered. (stages = 2 or stages = c(2,3,4,5) etc.)}

\item{results}{A dataframe - containing results for students. Columns titled "Stage", "nCorr", "nIncorr", "nDontKnow", "pctCorr", "pctIncorr" and "pctDontKnow" must be included. These columns should correspond to the stage, number correct, number incorrect, number don't know, percentage correct, percentage incorrect, percentage don't know answers respectively. Each student should have one entry (row) in this dataframe. This will be used to determine the number of students assessed in each stage}
}
\value{
A dataframe is returned containing the calculated response summaries A new column is generated for each stage considered
}
\description{
Response Summary Table
}
\examples{
fn_response_summary(stages = cnst$stages, results = dfResults)


}
