% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rm_cols_by_name.R
\name{fn_rm_cols_by_name}
\alias{fn_rm_cols_by_name}
\title{Remove Columns by Name}
\usage{
fn_rm_cols_by_name(input, items_to_rm)
}
\arguments{
\item{input}{The name of the data frame.}

\item{items_to_rm}{(Vector of numerics) The items to remove. Can be an empty vector, c()}
}
\value{
This function will return a dataframe with the selected columns removed
}
\description{
Removes specific columns by name from a dataframe. Additional functionality to handle scenarios where no columns need to be removed.
}
\examples{
fn_rm_cols_by_name(df_name, c(15,29,43)), fn_rm_cols_by_name(df_name, c())
}
