% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_startup_packages.R
\name{fn_startup_packages}
\alias{fn_startup_packages}
\title{Install and Load Required Packages}
\usage{
fn_startup_packages()
}
\value{
Invisibly returns TRUE if all packages are successfully loaded.
}
\description{
This function installs (if not already installed) and loads the necessary packages for the project.
}
\examples{
fn_startup_packages()
}
