% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_test_details.R
\name{fn_test_details}
\alias{fn_test_details}
\title{Test Details Parameters (Stage Specific)}
\usage{
fn_test_details(
  assessment = NULL,
  stages = NULL,
  programme = NULL,
  data_demog = NULL,
  results = NULL,
  results_absent = NULL
)
}
\arguments{
\item{assessment}{(OPTIONAL) A string - must define when used for IDS assessment for special conditions for that assessment type}

\item{stages}{A vector - denoting the stages to be considered. (stages = 2 or stages = c(2,3,4,5) etc.)}

\item{programme}{A string - the programme of the students (Must match what is in the "Programme" col in the data_demog dataframe (see below) eg "BMBS" etc.)}

\item{data_demog}{A dataframe - containing demographics data columns. Columns titled "Programme", "Stage" and "Year_Status" must be included. Make sure to use a df containing demographics data for all students and not one which has been filtered to include present students only}

\item{results}{A dataframe - containing results for students. A column titled "Stage" must be included. Each student should have one entry (row) in this dataframe. This will be used to determine the number of students assessed in each stage}

\item{results_absent}{A dataframe - containing results for absent students. A column titled "Stage" must be included. Each absent student should have one entry (row) in this dataframe. This will be used to determine the number of absent students who were not assessed in each stage}
}
\value{
A dataframe is returned (testDetails) containing the calculated test details. A new column is generated for each stage considered
}
\description{
Test Details Parameters (Stage Specific)
}
\examples{
fn_test_details(assessment = cnst$assessment_type, stages = cnst$stages, programme = cnst$programme, demog = dfDemog, results = dfRes, results_absent = dfResAbsent)
}
