% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnAnova.R
\name{fnAnova}
\alias{fnAnova}
\title{Perform ANOVA, save ANOVA results, adjusted means and observed means for different stages}
\usage{
fnAnova(
  demogData = NULL,
  colScore = NULL,
  varsAll = NULL,
  varsAnova = NULL,
  stages = NULL,
  reportObsMeanForAllVars = NULL
)
}
\arguments{
\item{demogData}{Dataframe containing demographics data columns (independent variables) and the student score (dependent variable) in another column and the relevant stage of each student row (if multiple stages are to be considered)}

\item{stages}{A vector denoting the stages to be considered. (stages = 2 or stages = c(2,3,4,5) etc.) If multiple stages are to be considered there must be a 'Stage' col in demogData}

\item{reportObsMeanForAllVars}{}
}
\description{
Perform ANOVA, save ANOVA results, adjusted means and observed means for different stages
}
