% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnCAPTPointBiserial.R
\name{fnCAPTPointBiserial}
\alias{fnCAPTPointBiserial}
\title{Calculate PtBis for CAPT Items}
\usage{
fnCAPTPointBiserial(
  data,
  colStudentID = "StudentID",
  colIsCorrect = "IsCorrect",
  colQuestionNumber = "QuestionNumber",
  significanceLevel = 0.05
)
}
\arguments{
\item{data}{(dataframe) Incoming long data eg a dataframe with cols StudentID, QuestionNumber and IsCorrect}

\item{colStudentID}{(default = "StudentID"), String The name of the column containing the Student ID number}

\item{colIsCorrect}{(default = "IsCorrect"), String. The name of the column containing the item score in each row}

\item{colQuestionNumber}{(default = "QuestionNumber"), String. The name of the column containing the question number in each row}

\item{significanceLevel}{(default = 0.05, 5%) Numeric. Set the significance level for each point biserial correlation}
}
\value{
returns a new dataframe storing the point biserial info for all input items
}
\description{
Calculate PtBis for CAPT Items
}
\examples{
df <- data.frame(


}
