% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnCAPTResponsesPivotWider.R
\name{fnCAPTResponsesPivotWider}
\alias{fnCAPTResponsesPivotWider}
\title{Pivoting CAPT responses to make wide dataframes from long}
\usage{
fnCAPTResponsesPivotWider(
  data,
  colIsCorrect = IsCorrect,
  colTopicCode = TopicCode,
  listTopicCodes = sort(unique(tab$blueprintCurr$TopicCode))
)
}
\arguments{
\item{data}{(dataframe) Incoming long data eg a dataframe with cols StudentID, TopicCode, IsCorrect}

\item{colIsCorrect}{(default = IsCorrect), String/variable. The name of the column containing the item scores in each row}

\item{colTopicCode}{(default = TopicCode), String/variable. The name of the column containing the topic codes in each row}

\item{listTopicCodes}{(default = sort(unique(tab$blueprintCurr$TopicCode))). List of all topic codes (either answered/unanswered)}
}
\value{
returns a new dataframe that has been pivoted to be wider
}
\description{
Pivoting CAPT responses to make wide dataframes from long
}
\examples{
df <- data.frame(
StudentID = c(
  "001",
  "001",
  "001",
  "001",
  "001",
  "002",
  "002",
  "002",
  "002",
  "002",
  "003",
  "003",
  "003",
  "003",
  "003"
),
TopicCode = c(
  "B",
  "C",
  "D",
  "E",
  "F",
  "A",
  "C",
  "D",
  "E",
  "F",
  "A",
  "B",
  "D",
  "E",
  "F"
),
IsCorrect = c(1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0)
)

pivotDf <- df \%>\%
  responsesPivot(
    colIsCorrect = IsCorrect,
    colTopicCode = TopicCode,
    listTopicCodes = c("A", "B", "C", "D", "E", "F")
  )


}
