% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnCreateSubfolder.R
\name{fnCreateSubfolder}
\alias{fnCreateSubfolder}
\title{Create a Subfolder in the Outputs Folder}
\usage{
fnCreateSubfolder(
  folderName = NULL,
  versionNumber = NULL,
  workingDirectory = NULL
)
}
\arguments{
\item{folderName}{(String) Name of the subfolder to create (if it doesn't exist already)}

\item{versionNumber}{(String) Version number appended to the Output folder name}

\item{workingDirectory}{(Path) Define the working directory, where the Output folder is}
}
\value{
The path of the created subfolder
}
\description{
Create a Subfolder in the Outputs Folder
}
\examples{
pathSubfolder <- fnCreateSubfolder(folderName = "Historic", versionNumber = cnst$verNum, workingDirectory = lFiles$pathWorkingDir)


}
