% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnGPNumToWord.R
\name{fnGPNumToWord}
\alias{fnGPNumToWord}
\title{Changes a whole number or integer value (up to 999) to a lower case character string,
with the option to convert to title case.}
\usage{
fnGPNumToWord(x, Capital=FALSE)
}
\arguments{
\item{x}{The value or object to convert.}

\item{Capital}{Default is FALSE. Option to create a title case character string with Capital=TRUE.}
}
\description{
This function takes an object, integer or whole number (up to a value of 999/"999") and will convert
it to a character string. Useful when including objects (e.g number of passing students) within report text.
Where the object or value is greater than 999/"999" and/or includes a decimal point the original string/
numeric value will be returned.
}
\examples{
x<-3
fnGPNumToWord(x)

fnGPNumToWord(12,TRUE)

fnGPNumToWord(999)

fnGPNumToWord(10.1)

studentsSick<-5
paste0("There were ",fnGPNumToWord(studentsSick)," students absent due to sickness.")

studentsECs<-8
paste0(fnGPNumToWord(studentsECs,TRUE)," students were absent with ECs.")

studentsECs<-0
paste0("There were ",fnGPNumToWord(studentsECs)," students with ECs.")

fnGPNumToWord(999,TRUE)

}
