% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnGPSchemeDetect.R
\name{fnGPSchemeDetect}
\alias{fnGPSchemeDetect}
\title{Function to detect and determine grading schemes .}
\usage{
fnGPSchemeDetect(Data, Force.Scheme)
}
\arguments{
\item{Data}{Data vector to from which to determine a Scheme.}

\item{Force.Scheme}{Optional. Used to specify or force a Scheme.}
}
\description{
Function to detect and determine grading schemes. fnGPSchemeDetect returns the scheme that has been determined from the data (SchemeDetect$Scheme), as well as a standardised data set (SchemeDetect$Data).

Can currently be used with: UBSE, USE, CIDK, CNINC, CNINC, CNINC, PF, PFE, UBLHE, Gender, Disability, Letters, Ethnicity
}
\note{
(DZ070917) Can't yet handle if two schemes account for the same amount of values; all possibles listed and warning displayed. First possible used to standardise the data.
(DZ070917) Doesn't check for values not in Table.Schemes; so rogue values in Data cause an error.
(DZ070917) Doesn't check scheme specified in Force.Scheme exists in Table.Schemes. Need check/warning for this.
}
\examples{

fnGPSchemeDetect()
fnGPSchemeDetect(Data=dataExample$Grade.UBSE)
fnGPSchemeDetect(Data=dataExample$Grade.UBSE, Force.Scheme = "UBSE")
fnGPSchemeDetect(Force.Scheme = "UBSE")
fnGPSchemeDetect(Data=dataExample)
fnGPSchemeDetect(Data=dataExample, Force.Scheme = "UBSE")
fnGPSchemeDetect(Data=c("U","S","E","U","S","E"))
fnGPSchemeDetect(Data=c(1,2,3,4,5))
fnGPSchemeDetect(Data=c(10.09,22.53,39.66,43.21,75.55))
fnGPSchemeDetect(Data=c(109,112,334,421,7655))
fnGPSchemeDetect(Data=c("male","female","m","f","f","MALE"))
fnGPSchemeDetect(Data=dataExample$Grade.UBSE, Force.Scheme = "Hermione") # Need a check for this
fnGPSchemeDetect(Data=c("U","S","B","U","S","W")) # Need to account for these

}
\author{
Dr Daniel Zahra, \email{daniel.zahra@plymouth.ac.uk}
}
