% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnHistoricStats.R
\name{fnHistoricStats}
\alias{fnHistoricStats}
\title{Historic Stats Table}
\usage{
fnHistoricStats(
  listOfConstants = NULL,
  listOfTables = NULL,
  assessment_type = NULL
)
}
\arguments{
\item{listOfConstants}{A list of contants. At least the following must be defined as list elements:
assessment
nStages
academicYear
cohort
stages
nItemsIncl
itemsExclCommaSep
testAngoffCorrVal
historicStatsComment}

\item{listOfTables}{A list of tables (dataframes). At least the following must be defined as list elements:
responseSummary
testDetails
reliability
gradeBoundaries}

\item{assessment_type}{The type of assessment (currently configured options: "ADK", "ADTK", "AMK")}
}
\value{
Returns a formatted dataframe of historic stats
}
\description{
Create a historic stats dataframe. NOTE: the inputs have to be of specific formats (mostly those generated by other functions as part of this package). If data has been saved in alternative formats, it is likely simpler to create the historic stats table manually.
}
\examples{
historicStats <- fnHistoricStats(listOfConstants = cnst, listOfTables = tab, assessment_type = cnst$assessment_type)


}
