% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnIDSScoreSummary.R
\name{fnIDSScoreSummary}
\alias{fnIDSScoreSummary}
\title{Create score summary tables for IDS analysis}
\usage{
fnIDSScoreSummary(lstOfDetails, dfRes, dfResAbsent, programmes, pctScoreTotal)
}
\arguments{
\item{lstOfDetails}{This variable should be a list containing as least the following defined variables: stage (stored as a numeric). This variable will be used to indentify the relevant students.}

\item{dfRes}{A dataframe containing the results for students who sat the exam}

\item{dfResAbsent}{A dataframe containing the results for students who did not sit the exam}

\item{programmes}{Which programmes to consider}

\item{pctScoreTotal}{Column name where the total percentage score is stored in dfRes}
}
\value{
A table of the score summaries is returned
}
\description{
Create score summary tables for IDS analysis
}
\examples{
fnIDSScoreSummary(cnst, dfRes, dfResAbsent, c("BDS"), pctScoreTotal)
}
