% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnPltGradeRange.R
\name{fnPltGradeRange}
\alias{fnPltGradeRange}
\title{Plot: Grade ranges for scores across multiple stages}
\usage{
fnPltGradeRange(
  data = NULL,
  grades = "UBSE",
  xLabels = NULL,
  axisTitles = NULL
)
}
\arguments{
\item{data}{A dataframe containing score data. For each group (stage) to be plotted, values should be added to a new column in the following order: First - minimum achieved score, then - grade boundaries in increasing order, finally - the maximum achieved score}

\item{grades}{The possible grades. Currently either "UBSE" (default) or "US}

\item{xLabels}{A vector (of strings or numerics). The labels to be added along the x-axis for each bar. Must be the same length as the number of columns (separate groups/stages) in the 'data' variable}

\item{axisTitles}{A vector (length 2), the x and the y axis titles (in that order)}
}
\value{
A plot of the grade ranges for scores across multiple stages is returned
}
\description{
Plot: Grade ranges for scores across multiple stages
}
\examples{

testDataUBSE <-
data.frame(
 # Order: Min, B bound., S bound., E bound., Max
 c(30, 4.2, 7.12, 38.23, 54),
 c(30, 22.6, 40.2, 62.84, 72),
 c(NA, 34.26, 45.45, 69.06, 78.25),
 c(20, 38, 52.2, 75.3, 85.2),
 row.names = c(
   'Minimum',
   'Borderline boundary',
   'Satisfactory boundary',
   'Excellent boundary',
   'Maximum'
 )
)

testDataUS <-
 data.frame(
   c(20.93, 25.1, 49.80),
   c(20.93, NA, 49.80),
   c(-41.8, 49.2, 68.8),
   row.names = c("Minimum", "Satisfactory boundary", "Maximum")
 )

fnPltGradeRange(
 data = testDataUBSE,
 grades = "UBSE",
 xLabels = c(1, 2, 3, 4),
 axisTitles = c("Stage", "Score (\%)")
)

fnPltGradeRange(
 data = testDataUS,
 grades = "US",
 xLabels = c(1, 2),
 axisTitles = c("Stage", "Score (\%)")
)


}
