% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnPltHistogram.R
\name{fnPltHistogram}
\alias{fnPltHistogram}
\title{Create a Histogram Plot}
\usage{
fnPltHistogram(data = NULL, binWidth = NULL, labelX = NULL, labelY = NULL)
}
\arguments{
\item{data}{The data to be plotted (usually a series of numbers - the series is converted to a dataframe as part of this function)}

\item{binWidth}{The bin width for plotting the histrogram}

\item{labelX}{The x-label of the accent}

\item{labelY}{The y-label of the accent}
}
\value{
A histrogram plot is returned
}
\description{
Create a Histogram Plot
}
\examples{

dataTest <- c(1,1,1,1,2,2,2,2,2,3,3,3,3,3,4,4,4,4)
fnPltHistogram(data = dataTest, binWidth = 0.5, labelX = "X Axis", labelY = "Y Axis")
}
