% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnPltHistogramScoresGrades.R
\name{fnPltHistogramScoresGradeFreq}
\alias{fnPltHistogramScoresGradeFreq}
\title{fnPltHistogramScoresGradeFreq}
\usage{
fnPltHistogramScoresGradeFreq(
  data = NULL,
  colScore = NULL,
  colGrade = NULL,
  scoreMax = NULL,
  scorePass = NULL
)
}
\arguments{
\item{data}{Dataframe containing the score and grade cols}

\item{colScore}{A string - name of the score col}

\item{colGrade}{A string - name of the grade col}

\item{scoreMax}{Numeric - The maximum possible score}

\item{scorePass}{Numeric - The pass score (added to plot with a line and annotation)}
}
\value{
The plot is returned
}
\description{
fnPltHistogramScoresGradeFreq
}
\examples{
fnPltHistogramScoresGradeFreq(data = dfRes,  colScore = "Score", colGrade = "Outcome",  scoreMax = scoreMax,  scorePass = passMark)


}
