% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnPltSave.R
\name{fnPltSave}
\alias{fnPltSave}
\title{Save a plot}
\usage{
fnPltSave(
  savePlot = NULL,
  plot = NULL,
  dimensions = c(15, 15, "cm"),
  filePath = NULL
)
}
\arguments{
\item{savePlot}{A boolean (TRUE/FALSE) determines whether the plot will be saved. Useful for controlling overall save behaviour.}

\item{plot}{The variable name of the plot - must exist in your environment}

\item{dimensions}{A vector containing the width, height and units of the dimensions of the saved image (in that order). If not set, defaults to 15cm x 15cm}

\item{filePath}{The file path where the plot is saved}
}
\value{
Nothing is explicitly returned, a plot is saved
}
\description{
Save a plot
}
\examples{
fnPltSave(savePlot = TRUE, plot = plt1, dimensions = c(15, 15, "cm"), filePath = glue('{some path}\\\\filename.jpg'))
}
