% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnReload.R
\name{fnReload}
\alias{fnReload}
\title{Reload the PsychometricGP package}
\usage{
fnReload()
}
\value{
Nothing is returned
}
\description{
Use this function to reload the current package (psychometricsGP). This function is useful when changes have been made and uploaded to the package repository. This function takes no arguments.
}
\examples{
fnReload()


}
