% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRmExclItems.R
\name{fnRmExclItems}
\alias{fnRmExclItems}
\title{Remove assessment items from a dataframe by specifying the column containing the item numbers/names and the list of items numbers/names to remove}
\usage{
fnRmExclItems(df = NULL, itemColName = NULL, itemsToRm = NULL)
}
\arguments{
\item{df}{The name of the data frame.}

\item{itemColName}{The name of the column containing the item numbers/names}

\item{itemsToRm}{A list of the items to remove}
}
\value{
A data frame with the specified items removed
}
\description{
Remove assessment items from a dataframe by specifying the column containing the item numbers/names and the list of items numbers/names to remove
}
\examples{
fnRmExclItems(df, itemColName, itemsToRm)
}
