% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRmNegMarking.R
\name{fnRmNegMarking}
\alias{fnRmNegMarking}
\title{Remove negative marking}
\usage{
fnRmNegMarking(scoresData = NULL, markOld = -0.25, markNew = 0)
}
\arguments{
\item{scoresData}{A single dataframe containing scores to be replaced. More columns may be present in the dataframe, but in all instances (including extra columns) all occurences of 'markOld' cells are replaced with 'markNew' cells}

\item{markOld}{(OPTIONAL) Defaults to '-0.25'. The cell value to be replaced (by the value of 'markNew')}

\item{markNew}{(OPTIONAL) Defaults to '0'. The cell value to go in the place of the cell value to replace ('markOld')}
}
\value{
The scoresData dataframe is returned with all the 'markOld' cells replaced with 'markNew'
}
\description{
Input a dataframe of scores where negative marking is to be removed. By default, all cells that are -0.25 are replaced with 0. Thought the exact replacement can be configured by defining the 'markOld' and 'markNew' optional variables.
}
\examples{
test <- fnRmNegMarking(scoresData = data.frame(c(1,1,-0.25), c(-0.25, 0, -0.25)))


}
