% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRmRowsByName.R
\name{fnRmRowsByName}
\alias{fnRmRowsByName}
\title{Removes specific columns by name from a dataframe. Additional functionality to handle scenarios where no columns need to be removed. 1D lists and 2D data frames are handled}
\usage{
fnRmRowsByName(input, itemsToRm)
}
\arguments{
\item{input}{The name of the data frame or list.}

\item{itemsToRm}{The items to remove. Can be an empty vector c()}
}
\value{
This function will return a dataframe with the selected rows removed
}
\description{
Removes specific columns by name from a dataframe. Additional functionality to handle scenarios where no columns need to be removed. 1D lists and 2D data frames are handled
}
\examples{
fnRmRowsByName(df, c(2)), fnRmRowsByName(df, c()), fnRmRowsByName(list, c(1,2,3))
}
