% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRoundTable.R
\name{fnRoundTable}
\alias{fnRoundTable}
\title{fnRoundTable Comprehensive Rounding of Dataframes for Reporting}
\usage{
fnRoundTable(
  data,
  rows = NULL,
  cols = NULL,
  row_decimals = NULL,
  col_decimals = NULL,
  rows2 = NULL,
  cols2 = NULL,
  row_decimals2 = NULL,
  col_decimals2 = NULL,
  rows_first = TRUE
)
}
\arguments{
\item{data}{A data frame to be rounded.}

\item{rows}{(Vector) Row indexes or row names to specify the rows for rounding.}

\item{cols}{(Vector) Column names to specify the columns for rounding.}

\item{row_decimals}{(Integer) The number of decimals for row rounding.}

\item{col_decimals}{(Integer) The number of decimals for column rounding.}

\item{rows2}{(Vector) Row indexes or row names for a second round of row rounding.}

\item{cols2}{(Vector) Column names for a second round of column rounding.}

\item{row_decimals2}{(Integer) The number of decimals for the second round of row rounding.}

\item{col_decimals2}{(Integer) The number of decimals for the second round of column rounding.}

\item{rows_first}{(Boolean) Indicating whether row rounding should be performed before column rounding.}
}
\value{
A modified data frame with rounded values. The cells are coerced to strings
}
\description{
Use this function to round dataframes. Define either rows or cols and the number of decimals to round each value row/col to. Rows/Cols can be given as vector of indexes or names. All cells in target rows/cols that cannot be converted to numerics are left as they were. By default rows are rounded first, followed by cols but this logic can be changed. Additionally, a second set of rows/cols can be defined so that they can be rounded to a different number of decimal places
}
\examples{
df <- data.frame(
  Name = c("John", "Jane", "Alice", "Bob"),
  Score1 = c("85.743", 90.123, "NA", 79.862),
  Score2 = c(92.567, "NA", 88.923, 95.369),
  Score3 = c(100.567, NA, 100.923, 902.329),
  Grade = c("A", "B", "C", "A"),
  row.names = c("Row1", "Row2", "Row3", "Row4")

# Example 1: Round specific rows (by row index) with specified decimals
rounded_rows_ny_index <- fnRoundTable(df, rows = c(1, 3), row_decimals = 1)

# Example 2: Round specific rows (by row name) with specified decimals
rounded_rows_by_name <- fnRoundTable(df, rows = c("Row1", "Row3"), row_decimals = 2)

# Example 3: Round specific columns (by col index) with specified decimals
rounded_cols_by_index <- fnRoundTable(df, cols = c(2, 3), col_decimals = 2)

# Example 4: Round specific columns (by col name) with specified decimals
rounded_cols_by_index <- fnRoundTable(df, cols = c("Score1", "Score3"), col_decimals = 2)

# Example 5: Perform row rounding before column rounding (DEFAULT)
rounded_rows_first <- fnRoundTable(df, rows = c(1), cols = c("Score1", "Score3"), row_decimals = 1, col_decimals = 4)

# Example 5: Perform col rounding before row rounding
rounded_rows_first <- fnRoundTable(df, rows = c(1), cols = c("Score1", "Score3"), row_decimals = 3, col_decimals = 1, rows_first = FALSE)

# Example 6: Perform a second round of row and column rounding
rounded_second_round <- fnRoundTable(df, rows = c(1), cols = c("Score1", "Score3"), row_decimals = 1, col_decimals = 2,
                                   rows2 = c(3), cols2 = c("Score2"), row_decimals2 = 2, col_decimals2 = 1)

}
\author{
Gergo Pinter
}
\keyword{dataframe}
\keyword{rounding}
\keyword{table}
