% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRoundTableOLD.R
\name{fnRoundTableOLD}
\alias{fnRoundTableOLD}
\title{fnRoundTable - Round Rows/Cols in a table for reports}
\usage{
fnRoundTableOLD(data, rows = NULL, cols = NULL, decimals, replaceNA = TRUE)
}
\arguments{
\item{data}{The dataframe to round}

\item{rows}{(Optional, Vector) Either numeric row indexes OR string row names}

\item{cols}{(Optional, Vector) Either numeric col indexes OR string col names}

\item{decimals}{Numeric, integer - the number of decimals to coerce in the chosen rows/cols}

\item{replaceNA}{(Optional, Default = TRUE) Boolean - decides whether all NAs in the data should be replaced with 0}
}
\value{
Returns the rounded input data
}
\description{
fnRoundTable - Round Rows/Cols in a table for reports
}
\examples{

# Create a sample data frame
df <- data.frame(
  Value1 = c(10.123, 15.789, 20.456, 25.678, 30.987),
  Value2 = c(5.678, 8.912, 12.345, 18.901, 23.456),
  Value3 = c(15.432, 19.876, 24.789, 29.123, 35.678),
  row.names = c("Row1", "Row2", "Row3", "Row4", "Row5")
)

# Round specific rows (by row name)
rounded_rows <- fnRoundTable(data = df, rows = c("Row2", "Row4"), decimals = 2)
print(rounded_rows)

# Round specific columns (by col name)
rounded_cols <- fnRoundTable(data = df, cols = c("Value2", "Value3"), decimals = 2)
print(rounded_cols)

#' # Round specific rows (by row index)
rounded_rows <- fnRoundTable(data = df, rows = c(1,2), decimals = 1)
print(rounded_rows)

# Round specific columns (by col index)
rounded_cols <- fnRoundTable(data = df, cols = c(4,5), decimals = 1)
print(rounded_cols)


}
