% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptAddParagraph.R
\name{fnRptAddParagraph}
\alias{fnRptAddParagraph}
\title{Add a new paragraph (1 line of vertical space) to a report}
\usage{
fnRptAddParagraph(rptVar = NULL)
}
\arguments{
\item{rptVar}{The variable name of the report}
}
\value{
Nothing is explicitly returned, rather a line of space is added to the configured report
}
\description{
Add a new paragraph (1 line of vertical space) to a report
}
\examples{
fnRptAddParagraph(rptVar = rpt$All)
}
