% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptAddPlot.R
\name{fnRptAddPlot}
\alias{fnRptAddPlot}
\title{Add a plot with accompanying caption}
\usage{
fnRptAddPlot(
  report = NULL,
  plot = NULL,
  plotCount = NULL,
  caption = NULL,
  stopTrailingLine = FALSE,
  dimensions = c(15.24, 10.16)
)
}
\arguments{
\item{report}{The variable name of the report}

\item{plot}{The variable name of the plot}

\item{plotCount}{The variable name in the document which stores the current number of plots already in the report (integer, usually named 'countPlt')}

\item{caption}{The caption to add (string)}

\item{stopTrailingLine}{TRUE/FALSE - decides whether to add an empty line below the plot (default is FALSE if undeclared)}

\item{dimensions}{(OPTIONAL) A vector defining the width and height (in cm) for the image. Defaults to 18cm wide and 12cm high}
}
\value{
The plot is added with an accompanying formatted caption to the configured report. The variable storing the number of plots in the report is incremented by 1 and a single trailing line is added following the caption (if requested)
}
\description{
Add a plot with accompanying caption
}
\examples{
fnRptAddPlot(report = rpt$All, plot = plot1, plotCount = countPlt, caption = "This caption", addTrailingLine = TRUE)
}
