% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptAddText.R
\name{fnRptAddText}
\alias{fnRptAddText}
\title{Add Text to a Report}
\usage{
fnRptAddText(report = NULL, text = NULL, style = "Normal")
}
\arguments{
\item{report}{The variable name of the report}

\item{text}{The text to add to the report}

\item{style}{The style of the text to add (DEFAULT = Normal)}
}
\value{
Nothing is explicitly returned, rather the configured text is added to the configured report
}
\description{
Add Text to a Report
}
\examples{
fnRptAddText(report = rpt$All, text = "Test text")
}
