% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptCaptionPlot.R
\name{fnRptCaptionPlot}
\alias{fnRptCaptionPlot}
\title{Add a new caption to accompany a plot and increment the value of the total number of plots.}
\usage{
fnRptCaptionPlot(report = NULL, plotCount = NULL, caption = NULL)
}
\arguments{
\item{report}{The variable name of the report}

\item{plotCount}{The variable name in the document which stores the current number of plots already in the report (integer, usually named 'countPlt')}

\item{caption}{The caption to add (string)}
}
\value{
The formatted caption is added to the configured report. The variable storing the number of plots in the reported is incremented by 1
}
\description{
Add a new caption to accompany a plot and increment the value of the total number of plots.
}
\examples{
fnRptCaptionPlot(report = rpt$All, plotCount = countPlt, caption = "This caption")
}
