% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptCaptionTab.R
\name{fnRptCaptionTab}
\alias{fnRptCaptionTab}
\title{Add a new caption to accompany a table and return the value of the total number of tables.}
\usage{
fnRptCaptionTab(report = NULL, tableCount = NULL, caption = NULL)
}
\arguments{
\item{report}{The variable name of the report}

\item{tableCount}{The variable name in the document which stores the current number of tables already in the report (integer, usually named 'countTab')}

\item{caption}{The caption to add (string)}
}
\value{
The new value for the total number of tables is returned. The output of this function should be saved to match whatever that varialbe is stored as in the main script. In addition to returning this, the formatted caption is also added the variable saving the report
}
\description{
Add a new caption to accompany a table and return the value of the total number of tables.
}
\examples{
countTab <- fnRptCaptionTab(rptVar = rpt$All, tableCount = countTab, caption = "This caption")
}
