% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptCaptionTable.R
\name{fnRptCaptionTable}
\alias{fnRptCaptionTable}
\title{Add a new caption to accompany a table and increment the value of the total number of tables.}
\usage{
fnRptCaptionTable(report = NULL, tableCount = NULL, caption = NULL)
}
\arguments{
\item{report}{The variable name of the report}

\item{tableCount}{The variable name in the document which stores the current number of variables already in the report (integer, usually named 'countTab')}

\item{caption}{The caption to add (string)}
}
\value{
The formatted caption is added to the configured report. The variable storing the number of tables in the reported is incremented by 1
}
\description{
Add a new caption to accompany a table and increment the value of the total number of tables.
}
\examples{
fnRptCaptionTable(rptVar = rpt$All, tableCount = countTab, caption = "This caption")
}
