% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptCoverSheet.R
\name{fnRptCoverSheet}
\alias{fnRptCoverSheet}
\title{Initialise the default cover page for a report}
\usage{
fnRptCoverSheet(rptVar = NULL, rptTitle = NULL, lstOfDetails = NULL)
}
\arguments{
\item{rptVar}{The variable name of the report}

\item{rptTitle}{This string will be the report title, shown prominently on the coverpage}

\item{lstOfDetails}{This variable should be a list containing as least the following defined variables: academicYear, programme, stage, module, assessmentNumber, assessmentDate. These variables will be input on the cover page.}
}
\value{
Nothin is explicitly returned, rather the cover page of the variable containing the report is populated with the required information.
}
\description{
Initialise the default cover page for a report
}
\examples{
fnRptCoverSheet(rptName, rptType, lstOfDetails)
}
