% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptDemographicAnalysis.R
\name{fnRptDemographicAnalysis}
\alias{fnRptDemographicAnalysis}
\title{ANOVA for Report}
\usage{
fnRptDemographicAnalysis(
  report = NULL,
  cnst = NULL,
  tab = NULL,
  tabDemog = NULL,
  tableCount = NULL
)
}
\arguments{
\item{report}{String - the variable name of the report which the analysis should be added to}

\item{cnst}{*** SHOULD CHANGE THIS - ONLY cnst$assessment is required}

\item{tab}{*** A list - to save any generated tables to}

\item{tabDemog}{*** A list - to save any generated demographics tables}

\item{tableCount}{Numeric (integer) - the current number of tables in the report (will be incremented as more tables are added)}
}
\value{
***
}
\description{
ANOVA for Report
}
\examples{
***
}
