% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptSectionTitle.R
\name{fnRptSectionTitle}
\alias{fnRptSectionTitle}
\title{Add a section title (level 2) to a report}
\usage{
fnRptSectionTitle(rptVar = NULL, sectionTitle = NULL)
}
\arguments{
\item{rptVar}{The variable name of the report}

\item{sectionTitle}{This string will be the added section title}
}
\value{
Nothing is explicitly returned, rather the desired section title is added to the report (this is one level higher than fnRptSectionHeading)
}
\description{
Add a section title (level 2) to a report
}
\examples{
fnRptSectionTitle(rptVar = rpt$All, sectionTitle = "Grade Distributions")
}
