% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptTabFormat.R
\name{fnRptTabFormat}
\alias{fnRptTabFormat}
\title{Format a table for presenting in a report}
\usage{
fnRptTabFormat(table, decimals, editColumnsNotRows, integerColumnsOrRows)
}
\arguments{
\item{table}{(Dataframe) The input data, which will be formatted}

\item{decimals}{(Integer, Numeric) The number of decimals to round values to}

\item{editColumnsNotRows}{(Boolean) If TRUE, the specified COLUMN numbers will be converted back to integers. If FALSE, the specified ROW numbers will be converted back to integers.}

\item{integerColumnsOrRows}{(OPTIONAL, Numeric, a list or single value eg: 1 or c(1,2,3)) The numbers of the COLUMNS or ROWS (depending on the editColumnsNotRows variable) which will be displayed as integers. If this variable is not defined, none of the columns or rows will be converted to integers.}
}
\value{
The formatted table is returned
}
\description{
This function's primary use is for report generation. Given an input dataframe containing only numbers (no text cells. numbers stored as text permissable) the entire dataframe is rounded to the desired number of decimal places. Then, depending on user selection, either the user-defined integer columns or rows are converted back to be integers (displaying no decimal places). The formatted table is returned.
}
\examples{
roundedTable <- fnRptTabFormat(table = tab$gradeBounds, decimals = 2, editColumnsNotRows = TRUE, integerColumnsOrRows = c(3, 4))


}
