% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnScoreStatistics.R
\name{fnScoreStatistics}
\alias{fnScoreStatistics}
\title{Score Statistics Table}
\usage{
fnScoreStatistics(
  assessmentType = NULL,
  listOfConstants = NULL,
  passMarks = data.frame(stringsAsFactors = FALSE)
)
}
\arguments{
\item{assessmentType}{A string. The type of assessment being analysed. This will determine the selection process for finding and returning items to review}

\item{listOfConstants}{A list of contants. At least the following must be defined as list elements: assessment (Eg "ADK38"), nPresent (the number of students who sat the assessment), nItems (the original number of items), nItemsExcl (the number of items excluded), itemsExclStr (a string of the specific item numbers removed)}

\item{passMarks}{(OPTIONAL - Required for ADK, ADTK) A data frame containing the relevant passmarks. Usually a single row with cols for each stage (named "Stage 1", "Stage 2" etc.) The code of this function can be changed to match the specific formats that this table can take for different assessmentTypes.}
}
\value{
A dataframe of the score statistics (the main first table added to the assessment report)
}
\description{
Score Statistics Table
}
\examples{
tab$scoreStatistics <- fnScoreStatistics(
assessmentType = "ADK",
listOfConstants = cnst,
passMarks =tab$passMarks)


}
\author{
Gergo Pinter, \email{gergo.pinter@plymouth.ac.uk}, \url{https://www.gergo.io}
}
