% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnStartFolders.R
\name{fnStartFolders}
\alias{fnStartFolders}
\title{Create Output, Pulse and Images folders.}
\usage{
fnStartFolders(
  workingDirectory = NULL,
  versionNumber = NULL,
  createOutputFolder = TRUE,
  createPulseSubfolder = TRUE,
  createImagesFolder = TRUE
)
}
\arguments{
\item{workingDirectory}{Define the working directory, where save folders will be created}

\item{versionNumber}{Set a version number - appended to save folders and file names}

\item{createOutputFolder}{(TRUE/FALSE/NULL) Create an outputs folder in the working directory (with appended version number) if TRUE}

\item{createImagesFolder}{(TRUE/FALSE/NULL) Create an Images folder in the Output folder if TRUE}

\item{createPulseFolder}{(TRUE/FALSE/NULL) Create a folder called Pulse in the Output folder if TRUE}
}
\value{
A list of the paths to the created folders (for Outputs, Images and Pulse Files) is returned.
}
\description{
Create Output, Pulse and Images folders.
}
\examples{
listOfPaths <- fnStartFolders(workingDirectory = lFiles$pathWorkingDir #or getwd(), versionNumber = "v1", createOutputFolder = TRUE, createPulseSubfolder = FALSE, createImagesFolder = TRUE)
}
