% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnStartPackages.R
\name{fnStartPackages}
\alias{fnStartPackages}
\title{Run functions to install and load required packages}
\usage{
fnStartPackages()
}
\value{
Required packages are installed and loaded if they are not already.
}
\description{
Run functions to install and load required packages
}
\examples{
fnStartPackages()
}
