% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTestAngoffCorr.R
\name{fnTestAngoffCorr}
\alias{fnTestAngoffCorr}
\title{Test Angoff Correlation}
\usage{
fnTestAngoffCorr(angoffScores = NULL, testFacility = NULL)
}
\arguments{
\item{angoffScores}{A vector (numeric elements) - a vector of all the included angoff scores for each item. (Must be the same length as testFacility)}

\item{testFacility}{A vector (numeric elements) - a vector of all the included item scores (facility). Usually, the highest stage is used. (Must be the same length as angoffScores)}
}
\value{
A list of constants is returned. These include the correlation coefficient and the bounds of the 95 confidence interval.
}
\description{
Test Angoff Correlation
}
\examples{
cnst <- append(
cnst,
 fnTestAngoffCorr(
 angoffScores = fnRmRowsByName(dfAngoff$Score, cnst$itemsExcl),
 testFacility = tab$itemAnalysis[["Stage 5 Facility"]]
 )
 )


}
