% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTestDetails.R
\name{fnTestDetails}
\alias{fnTestDetails}
\title{Test Details Parameters (Stage Specific)}
\usage{
fnTestDetails(
  stages = NULL,
  demogData = NULL,
  results = NULL,
  resultsAbsent = NULL
)
}
\arguments{
\item{stages}{A vector - denoting the stages to be considered. (stages = 2 or stages = c(2,3,4,5) etc.)}

\item{demogData}{A dataframe - containing demographics data columns. Columns titled "Programme", "Stage" and "Year_Status" must be included}

\item{results}{A dataframe - containing results for students. A column titled "Stage" must be included. Each student should have one entry (row) in this dataframe. This will be used to determine the number of students assessed in each stage}

\item{resultsAbsent}{A dataframe - containing results for absent students. A column titled "Stage" must be included. Each absent student should have one entry (row) in this dataframe. This will be used to determine the number of absent students who were not assessed in each stage}
}
\value{
A dataframe is returned (testDetails) containing the calculated test details. A new column is generated for each stage considered
}
\description{
Test Details Parameters (Stage Specific)
}
\examples{
fnTestDetails(stages = cnst$stages, demog = dfDemog, results = dfRes, resultsAbsent = dfResAbsent)
}
