% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTestRetest.R
\name{fnTestRetest}
\alias{fnTestRetest}
\title{Test Retest Dataframe Creation}
\usage{
fnTestRetest(
  assessment = NULL,
  assessmentPrev = NULL,
  results = NULL,
  resultsPrevious = NULL
)
}
\arguments{
\item{assessment}{A string - the current assessment being analysed (eg ADK38)}

\item{assessmentPrev}{A string - the previous assessment (prior to the one being analysed - eg ADK37)}

\item{results}{A dataframe - containing results data for the current assessment with at least the following columns: "StudentID", "Stage", assessement score (of the form "assessement_Score" where assessment matches the variable defined in the "assessment" parameter) and assessment grade (of the form "assessment_Grade" where assessment matched the variable defined in the "assessment" parameter)}

\item{resultsPrevious}{A dataframe - containing results data for the previous assessment, in a similar fashion to the "results" dataframe}
}
\value{
A list of dataframes is returned. The results dataframe contains the student ID, stage and current and previous assessment grades and scores. The stagesResults list of dataframes contains the same data as the results dataframe with additional stage separation
}
\description{
Test Retest Dataframe Creation
}
\examples{
testRetest <- append(testRetest, fnTestRetest(assessment = cnst$assessment, assessmentPrev = cnst$assessmentPrev, results = dfRes, resultsPrevious = dfPrevTestData)
}
