% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTestRetestAnalysis.R
\name{fnTestRetestAnalysis}
\alias{fnTestRetestAnalysis}
\title{Test Retest Data Analysis}
\usage{
fnTestRetestAnalysis(lstOfDetails, dataAnalysis, dataRaw)
}
\arguments{
\item{lstOfDetails}{This variable should be a list containing as least the following defined variables: assessment (the current assessment number - eg: PT36) and assessmentPrev (the previous assessment number - eg: PT35)}

\item{dataAnalysis}{A list - the name of the variable storing all the test retest data analysis results}

\item{dataRaw}{A dataframe - all the raw test retest data. The following columns must be present:}
}
\value{
todo
}
\description{
Test Retest Data Analysis
}
\examples{
todo
}
