% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTestRetestMatrix.R
\name{fnTestRetestMatrix}
\alias{fnTestRetestMatrix}
\title{Test Retest Data Matrix Creation}
\usage{
fnTestRetestMatrix(
  stage = NULL,
  assessment = NULL,
  assessment_prev = NULL,
  dataRaw = NULL
)
}
\arguments{
\item{stage}{A number or string (OPTIONAL). If a stage is defined, the stage will be appended to the saved list items.  If a stage is not defined, no stage info will be appended.}

\item{assessment}{A string - the current assessment being analysed (eg ADK38)}

\item{assessment_prev}{A string - the previous assessment (prior to the one being analysed - eg ADK37)}

\item{dataRaw}{A dataframe - all the raw test retest data. When using this function for a specific stage, the input raw data must be for that stage only. No stage separation is performed as part of the function. The following columns must be present: "StudentID" and also the test grades for the current assessement and the previous assessment in the form: "####_Grade" (eg for analysis of PT36: "PT35_Grade", "PT36_Grade")}
}
\value{
A list is returned. The list contains all the test retest analysis. Analysis is specific to whatever data is input. Variable names will contain the given stage if it is specified. If saving this data to an existing list, use the append function (as shown in the example)
}
\description{
Test Retest Data Matrix Creation
}
\examples{
testRetest <- append(testRetest, fnTestRetestMatrix(stage = 1, assessment = cnst$assessment, assessment_prev = cnst$assessment_prev, dataRaw = testRetest$resultsStages[["Stage 1"]]))
}
