% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTimestamp.R
\name{fnTimestamp}
\alias{fnTimestamp}
\title{Return a custom formatted timestamp}
\usage{
fnTimestamp()
}
\value{
Returns a timestamp string
}
\description{
Return a custom formatted timestamp
}
\examples{
fnTimestamp()
}
