% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_process_scores.R
\name{process_scores}
\alias{process_scores}
\title{Process Scores from Responses}
\usage{
process_scores(
  responses,
  demog,
  answer_key,
  dont_know_value,
  items_excluded,
  stages
)
}
\arguments{
\item{responses}{A data frame containing student responses, including the columns
for `StudentID`, `Stage`, and the responses for each question.}

\item{demog}{A data frame containing demographic information, including
`StudentID` and `Stage`.}

\item{answer_key}{A data frame containing the correct answers for each question,
typically with a column named `CorrectAnswer`.}

\item{dont_know_value}{A value indicating the response for "don't know" (e.g., 0 or 6).}

\item{items_excluded}{A character vector of item names to be excluded from the scores.}

\item{stages}{A numeric vector indicating the stages used for separating the scores.}
}
\value{
A list containing several data frames:
  \item{scores_all_neg}{Data frame of scores with negative marking applied.}
  \item{scores_ids_all_neg}{Data frame of scores with IDs and negative marking applied.}
  \item{scores_neg}{Data frame of scores after removing excluded items and negative marking.}
  \item{scores_ids_neg}{Data frame of scores with IDs after removing excluded items and negative marking.}
  \item{scores}{Data frame of scores after removing negative marking.}
  \item{scores_ids}{Data frame of scores with IDs after removing negative marking.}
  \item{scores_all}{Data frame of all scores after removing negative marking.}
  \item{scores_ids_all}{Data frame of all scores with IDs after removing negative marking.}
  \item{stages_scores_all_neg}{List of data frames, each containing scores for a specific stage with negative marking applied.}
  \item{stages_scores_all}{List of data frames, each containing scores for a specific stage after removing negative marking.}
  \item{stages_scores_neg}{List of data frames, each containing scores for a specific stage after removing excluded items and negative marking.}
  \item{stages_scores}{List of data frames, each containing scores for a specific stage after removing excluded items.}
}
\description{
This function processes student responses to generate scores based on an answer key,
handles negative marking, removes excluded items, and separates scores by stage.
}
\examples{
# Example usage of process_scores
results <- process_scores(responses = df_responses,
                           demog = df_demographics,
                           answer_key = df_answer_key,
                           dont_know_value = 6,
                           items_excluded = c("Item1", "Item2"),
                           stages = 1:3)

}
