% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeGP.R
\name{themeGP}
\alias{themeGP}
\title{A custom ggplot theme for PSMD.
Some edits to come by GP. Most work by DZ}
\usage{
theme_psmd()
}
\description{
A custom ggplot theme for PSMD.
}
\examples{
 ## Uses data from Data.Example in PSMD.Psychometrics
## Requires ggplot2 to be loaded

ggplot(data=Data.Example, aes(x=Grade.UBSE)) +
  geom_bar(stat="count", position=position_dodge()) +
  ggtitle("Basic Bar-Plot") +
  theme_psmd()

}
\author{
Dr Daniel Zahra, \email{daniel.zahra@plymouth.ac.uk}
}
