% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_historic_stats.R
\name{fn_historic_stats}
\alias{fn_historic_stats}
\title{Historic Stats Table}
\usage{
fn_historic_stats(exam_vars = NULL, l_tables = NULL)
}
\arguments{
\item{exam_vars}{A list of contants. At least the following must be defined as list elements:
exam
exam_type
n_stages
academic_year
cohort
stages
n_items_incl
items_excl_comma_sep
testAngoffCorrVal
historic_stats_comment}

\item{l_tables}{A list of tables (dataframes). At least the following must be defined as list elements:
responseSummary
testDetails
reliability
gradeBoundaries}
}
\value{
Returns a formatted dataframe of historic stats
}
\description{
Create a historic stats dataframe. NOTE: the inputs have to be of specific formats (mostly those generated by other functions as part of this package). If data has been saved in al_tablesernative formats, it is likely simpler to create the historic stats table manually.
}
\examples{
historic_stats <- fn_historic_stats(exam_vars = cnst, l_tables = tab)


}
