% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plt_borderline_regression.R
\name{fn_plt_borderline_regression}
\alias{fn_plt_borderline_regression}
\title{Plot of the Borderline Regression method.}
\source{
Written by Martin Roberts (psychometrics@plymouth.ac.uk)
}
\usage{
fn_plt_borderline_regression(x, grade_col="Grade", score_col="Score")
}
\arguments{
\item{x}{A data frame containing the global grade and score data.}

\item{grade_col}{Name of the column containing the global grades.}

\item{score_col}{Name of the column containing the assessment scores.}
}
\value{
A list containing:
\item{plt_borderline_regression}{A ggplot object which can be saved or customised as needed}
\item{pass_mark}{The calculated pass mark from the borderline regression}
Additionally the linear regression model (b_reg_line) is passed to the global environment.
}
\description{
fn_plt_borderline_regression creates a scatterplot assessment scores against global grades.
A regression line is added to the plot and the intersection of this line with the 'borderline group' is used to calculate a pass mark.
The main output from fn_plt_borderline_regression is a list containing a ggplot object and the pass mark.
Additionally the function passes the linear regression model into the global environment (see Examples).
}
\note{
This function uses \code{\link[ggplot2]{ggplot2}} for the underlying graphics,
and uses the str_wrap function from the \code{\link[stringr]{stringr}} package to wrap the x axis labels.
}
\examples{
library("ggplot2")
data <- data.frame(Grade = factor(c(rep("Unsatisfactory",5),rep("Borderline",10),rep("Low Satisfactory",20),
       rep("High Satisfactory",25),rep("Excellent",10)),ordered=TRUE,
       levels = c("Unsatisfactory","Borderline","Low Satisfactory","High Satisfactory","Excellent")),
       Score = c(5,8:10,10, 10:14,11:15, 12:16,13:17,13:17,14:18, 13:17,14:18,15:19,16:20,17:21, 18:22,20:23,25),
       Grade4 = factor(c(rep("Unsatisfactory",5),rep("Borderline",20),rep("Satisfactory",32),rep("Excellent",13)),
       ordered = TRUE, levels = c("Unsatisfactory","Borderline","Satisfactory","Excellent")),
       Score4 = c(5,6,8:10, 9:13,10:14, 10:14,11:15,11:15,12:16, 12:16,13:17,13:17,14:18,14:18,15:20, 18:20,20))

results <- fn_plt_borderline_regression(data)
results$plt_borderline_regression  # The plot
results$pass_mark  # The pass mark
b_reg_line  # Details of the regression model
summary(b_reg_line)   # Summary of the regression model
car::outlierTest(b_reg_line)   # Test for significant outliers in the model

results2 <- fn_plt_borderline_regression(data, "Grade4", "Score4")


}
