% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_point_biserial_significant.R
\name{fn_point_biserial_significant}
\alias{fn_point_biserial_significant}
\title{Calculate PtBis and Statistical Significance for Test Items}
\usage{
fn_point_biserial_significant(
  data,
  assessment = "Test",
  col_student_id = "StudentID",
  col_is_correct = "IsCorrect",
  col_item_id = "ItemID",
  significance_level = 0.05
)
}
\arguments{
\item{data}{(dataframe) Incoming long data eg a dataframe with cols student_id, item_id, and is_correct}

\item{assessment}{(String) The name of the test.}

\item{col_student_id}{(default = "StudentID"), String The name of the column containing the Student ID number}

\item{col_is_correct}{(default = "IsCorrect"), String. The name of the column containing the item score in each row}

\item{col_item_id}{(default = "ItemID"), String. The name of the column containing the ItemID in each row}

\item{significance_level}{(default = 0.05) Numeric. Set the significance level for each point biserial correlation}
}
\value{
returns a new dataframe storing the point biserial info for all input items
}
\description{
Calculate PtBis and Statistical Significance for Test Items
}
\examples{
df <- data.frame(
  StudentID = rep(1:3, each = 6),
  ItemID = rep(c("A", "B", "C", "D", "E", "F"), 3),
  IsCorrect = c(1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0)
)
df |> fn_point_biserial_significant()
}
