% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnItemAnalysis.R
\name{fnItemAnalysis}
\alias{fnItemAnalysis}
\title{Item Analysis}
\usage{
fnItemAnalysis(stages = NULL, scores = NULL, excludedItems = NULL)
}
\arguments{
\item{stages}{***}

\item{scores}{*** Stage separated (each list item should be stage1 etc. with all items showing non-negatively marked item (col) scores for each student (row) in that stage lResults$stagesScoresAll}

\item{excludedItems}{***}
}
\value{
***
}
\description{
Item Analysis
}
\examples{
***


}
