% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnItemReview.R
\name{fnItemReview}
\alias{fnItemReview}
\title{Detect Items to Review}
\usage{
fnItemReview(assessmentType = NULL, itemAnalysis = NULL)
}
\arguments{
\item{assessmentType}{A string. The type of assessment being analysed. This will determine the selection process for finding and returning items to review}

\item{itemAnalysis}{A dataframe. This dataframe must contain at least the following columns: "Item" (the items number), "Stage 5 Facility" (for ADK), "Stage 5 vs 2 Growth" (for ADK), "Stage 5 PtBis" (for ADK)}
}
\value{
A list of the detected items to review is returned.
}
\description{
Detect Items to Review
}
\examples{
#For ADK:
 cnst <- append(
 cnst,
  fnItemReview(assessmentType = cnst$assessmentType, itemAnalysis = tab$itemAnalysis)
  )


}
