% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnItemReview.R
\name{fnItemReview}
\alias{fnItemReview}
\title{Detect Items to Review}
\usage{
fnItemReview(assessmentType = NULL, itemAnalysisData = NULL, testInYear = NULL)
}
\arguments{
\item{assessmentType}{A string. The type of assessment being analysed. This will determine the selection process for finding and returning items to review}

\item{itemAnalysisData}{A dataframe. This dataframe must contain at least the following columns: "Item" (the items number), "Stage 5 Facility" (for ADK), "Stage 5 vs 2 Growth" (for ADK), "Stage 5 PtBis" (for ADK) OR "Stage 3 Facility" (for ADTK), "Stage 3 vs 2 Growth" (for ADTK), "Stage 3 PtBis" (for ADTK) OR "Stage 2 Facility" (for PAPT), "Stage 2 vs 1 Growth" (for PAPT), "Stage 2 PtBis" (for PAPT) OR "Facility" (for IDS/Y1KT), "PtBis" (for IDS/Y1KY)}

\item{testInYear}{(OPTIONAL unless AMK) - Required for AMK assessments so the stage to use for item review can be determined}
}
\value{
A list of the detected items to review is returned.
}
\description{
Detect Items to Review
}
\examples{
#For ADK/ADTK/PAPT/IDS/Y1KT/AMK/AKT:
 cnst <- append(
 cnst,
  fnItemReview(assessmentType = cnst$assessmentType, itemAnalysisData = tab$itemAnalysis)
  )

# For AMK:
 cnst <- append(
 cnst,
  fnItemReview(assessmentType = cnst$assessmentType, itemAnalysisData = tab$itemAnalysis, testInYear = cnst$testInYear)
  )


}
