% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnListsFromVector.R
\name{fnListsFromVector}
\alias{fnListsFromVector}
\title{Create Empty Lists From a Vector of Strings}
\usage{
fnListsFromVector(listsToCreate = NULL)
}
\arguments{
\item{listsToCreate}{A vector of strings.}
}
\value{
Creates the required lists from the input vector in the main (global) environment
}
\description{
This function requires an input of a vector of strings. Each element of the vector is checked. If the element does not already exist as a variable, then a new variable (with the name of the vector element) is created. The variable is assigned as an empty list.
}
\examples{
fnListsFromVector(listsToCreate = c("itemAnalysis", "lReliab", "lResults", "plt", "rpt", "rpt$tabAll", "tab", "tabPulse", "testRetest", "testAngoff", "tabDemog"))


}
