% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnPltHistoricStatsMultiple.R
\name{fnPltHistoricStatsMultiple}
\alias{fnPltHistoricStatsMultiple}
\title{Title}
\usage{
fnPltHistoricStatsMultiple(listOfCurrentPlots, listOfData, listOfStats)
}
\arguments{
\item{listOfCurrentPlots}{***}

\item{listOfData}{***}

\item{listOfStats}{***}
}
\value{
****
}
\description{
Title
}
\examples{
****
}
