% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnReliability.R
\name{fnReliability}
\alias{fnReliability}
\title{Test Reliability Parameters (Stage Specific)}
\usage{
fnReliability(stages = NULL, scoresData = NULL, resultsData = NULL)
}
\arguments{
\item{stages}{Vector containing all the stages being considered}

\item{scoresData}{List of dataframes with students scores for each stage. Each data frame in the list should be named "Stage #" where # is the relevant stage. In each dataframe the columns should be the items and the rows should be the index of each student (actual value unimportant). Each data point should represent the score for a particular student on a particular item. All students with NA scores and excluded items should be removed prior to use in this function}

\item{resultsData}{A dataframe containing student results. This variable must contain a "Stage" column in which the stage of each student is entered in a new row. This is used to add the number of students assessed in each Stage to the resultant export of reliability data.}
}
\value{
A dataframe containing the reliability data for each considered stage is returned
}
\description{
Test Reliability Parameters (Stage Specific)
}
\examples{
fnReliability(stages = cnst$stages, scoresData = dfResStagesOnlyScores)
}
