% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRmColsByName.R
\name{fnRmColsByName}
\alias{fnRmColsByName}
\title{Removes specific columns by name from a dataframe}
\usage{
fnRmColsByName(df, itemsToRm)
}
\arguments{
\item{df}{The name of the data frame.}

\item{itemsToRm}{The}
}
\value{
This function will return a dataframe with the selected columns removed
}
\description{
Removes specific columns by name from a dataframe
}
\examples{
fnRmColsByName(dfName, c(15,29,43))
}
