% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRmRowsByName.R
\name{fnRmRowsByName}
\alias{fnRmRowsByName}
\title{Remove Rows by Name}
\usage{
fnRmRowsByName(input, itemsToRm)
}
\arguments{
\item{input}{The name of the data frame or list.}

\item{itemsToRm}{(A vector of numerics) The items to remove. Can be an empty vector, c()}
}
\value{
This function will return a dataframe with the selected rows removed
}
\description{
Removes specific rows by name from a dataframe or list. Additional functionality to handle scenarios where no columns need to be removed. Works for 1D lists and 2D data frames.
}
\examples{
fnRmRowsByName(dfName, c(2)), fnRmRowsByName(dfName, c()), fnRmRowsByName(list, c(1,2,3))
}
