% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptAddPlot.R, R/fnRptAddTable.R
\name{fnRptAddPlot}
\alias{fnRptAddPlot}
\title{Add a plot with accompanying caption}
\usage{
fnRptAddPlot(
  report = NULL,
  table = NULL,
  tableCount = NULL,
  caption = NULL,
  addTrailingLine = NULL
)

fnRptAddPlot(
  report = NULL,
  table = NULL,
  tableCount = NULL,
  caption = NULL,
  addTrailingLine = NULL
)
}
\arguments{
\item{report}{The variable name of the report}

\item{table}{The variable name of the table}

\item{tableCount}{The variable name in the document which stores the current number of tables already in the report (integer, usually named 'countTab')}

\item{caption}{The caption to add (string)}

\item{addTrailingLine}{TRUE/FALSE}

\item{plot}{The variable name of the plot}

\item{plotCount}{The variable name in the document which stores the current number of plots already in the report (integer, usually named 'countPlt')}
}
\value{
The plot is added with an accompanying formatted caption to the configured report. The variable storing the number of plots in the report is incremented by 1 and a single trailing line is added following the caption (if requested)

The table is added with an accompanying formatted caption to the configured report. The variable storing the number of tables in the report is incremented by 1 and a single trailing line is added following the caption (if requested)
}
\description{
Add a plot with accompanying caption

Add a table with accompanying caption
}
\examples{
fnRptAddPlot(report = rpt$All, plot = plot1, plotCount = countPlt, caption = "This caption", addTrailingLine = TRUE)
fnRptAddTable(report = rpt$All, table = table1, tableCount = countTab, caption = "This caption", addTrailingLine = TRUE)
}
