% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptAddPlot.R
\name{fnRptAddPlot}
\alias{fnRptAddPlot}
\title{Add a plot with accompanying caption}
\usage{
fnRptAddPlot(
  report = NULL,
  plot = NULL,
  plotCount = NULL,
  caption = NULL,
  addTrailingLine = NULL
)
}
\arguments{
\item{report}{The variable name of the report}

\item{plot}{The variable name of the plot}

\item{plotCount}{The variable name in the document which stores the current number of plots already in the report (integer, usually named 'countPlt')}

\item{caption}{The caption to add (string)}

\item{addTrailingLine}{TRUE/FALSE}
}
\value{
The plot is added with an accompanying formatted caption to the configured report. The variable storing the number of plots in the report is incremented by 1 and a single trailing line is added following the caption (if requested)
}
\description{
Add a plot with accompanying caption
}
\examples{
fnRptCaptionPlot(report = rpt$All, plot = plot1, plotCount = countPlt, caption = "This caption")
}
