% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptAddTable.R
\name{fnRptAddTable}
\alias{fnRptAddTable}
\title{Add a table with accompanying caption}
\usage{
fnRptAddTable(
  report = NULL,
  table = NULL,
  tableCount = NULL,
  caption = NULL,
  stopTrailingLine = FALSE,
  stopFlextableConversion = FALSE
)
}
\arguments{
\item{report}{The variable name of the report}

\item{table}{The variable name of the table}

\item{tableCount}{The variable name in the document which stores the current number of tables already in the report (integer, usually named 'countTab')}

\item{caption}{The caption to add (string)}

\item{stopTrailingLine}{TRUE/FALSE - decides whether to add an empty line below the table (default is FALSE if undeclared)}

\item{stopFlextableConversion}{TRUE/FALSE - decides whether to convert the input table to a flextable. If the table has already been converted to a flextable prior to its use in the function. Set stopFlextableCoversion to FALSE. Otherwise, if a dataframe is input to this function stopFlextableCoversion can be left undeclared (default is FALSE if undeclared)}
}
\value{
The table is added with an accompanying formatted caption to the configured report. The variable storing the number of tables in the report is incremented by 1 and a single trailing line is added following the caption (if requested)
}
\description{
Add a table with accompanying caption
}
\examples{
fnRptAddTable(report = rpt$All, table = table1, tableCount = countTab, caption = "This caption", addTrailingLine = TRUE)
}
