% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptAddText.R
\name{fnRptAddText}
\alias{fnRptAddText}
\title{Add Text to a Report}
\usage{
fnRptAddText(rptVar = NULL, text = NULL)
}
\arguments{
\item{rptVar}{The variable name of the report}

\item{text}{The text to add to the report}
}
\value{
Nothing is explicitly returned, rather the configured text is added to the configured report
}
\description{
Add Text to a Report
}
\examples{
fnRptAddText(rptVar = rpt$All, text = "Test text")
}
