% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptCaptionPlot.R
\name{fnRptCaptionPlot}
\alias{fnRptCaptionPlot}
\title{Add a new caption to accompany a plot}
\usage{
fnRptCaptionPlot(rptVar = NULL, caption = NULL, plotCount = NULL)
}
\arguments{
\item{rptVar}{The variable name of the report}

\item{caption}{The caption to add (string)}

\item{plotCount}{The current number of plots already in the report (integer)}
}
\value{
Nothing is explicitly returned, rather a formatted caption is added to the report
}
\description{
Add a new caption to accompany a plot
}
\examples{
fnRptCaptionPlot(rptVar = rpt$All, caption = "This caption", plotCount = 2)
}
