% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptCoverSheet.R
\name{fnRptCoverSheet}
\alias{fnRptCoverSheet}
\title{Initialise the default cover page for a report}
\usage{
fnRptCoverSheet(reportVar = NULL, reportTitle = NULL, listOfDetails = NULL)
}
\arguments{
\item{reportVar}{The variable name of the report}

\item{reportTitle}{This string will be the report title, shown prominently on the coverpage}

\item{listOfDetails}{This variable should be a list containing as least the following defined variables: academicYear, programme (OPTIONAL: programme is overridden by programmeLong so long as programmeLong is defined) stage/stages (the function automatically handles either scenario with the default being to use stage), module, assessmentNumber, assessmentDate. These variables will be input on the cover page.}
}
\value{
Nothing is explicitly returned, rather the cover page of the variable containing the report is populated with the required information.
}
\description{
Initialise the default cover page for a report
}
\examples{
fnRptCoverSheet(reportVar, reportTitle, listOfDetails)
}
