% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptDemographicAnalysis.R
\name{fnRptDemographicAnalysis}
\alias{fnRptDemographicAnalysis}
\title{ANOVA for Report}
\usage{
fnRptDemographicAnalysis(
  report = NULL,
  cnst = NULL,
  tab = NULL,
  tabDemog = NULL,
  tableCount = NULL
)
}
\arguments{
\item{report}{***}

\item{cnst}{***}

\item{tab}{***}

\item{tabDemog}{***}

\item{tableCount}{***}
}
\value{
***
}
\description{
ANOVA for Report
}
\examples{
***
}
