% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptSectionHeading.R
\name{fnRptSectionHeading}
\alias{fnRptSectionHeading}
\title{Add a section heading to a report}
\usage{
fnRptSectionHeading(rptVar = NULL, headingTitle = NULL)
}
\arguments{
\item{rptVar}{The variable name of the report}

\item{headingTitle}{This string will be the added section heading}
}
\value{
Nothing is explicitly returned, rather the desired section heading is added to the report
}
\description{
Add a section heading to a report
}
\examples{
fnRptSectionHeading(rptVar = rpt$All, headingTitle = "Grade Distributions")
}
