% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnRptSectionHeading.R, R/fnRptSectionTitle.R
\name{fnRptSectionHeading}
\alias{fnRptSectionHeading}
\title{Add a section heading (level 3) to a report}
\usage{
fnRptSectionHeading(rptVar = NULL, sectionTitle = NULL)

fnRptSectionHeading(rptVar = NULL, sectionTitle = NULL)
}
\arguments{
\item{rptVar}{The variable name of the report}

\item{sectionTitle}{This string will be the added section title}

\item{headingTitle}{This string will be the added section heading}
}
\value{
Nothing is explicitly returned, rather the desired section heading is added to the report (this is one level lower than fnRptSectionTitle)

Nothing is explicitly returned, rather the desired section title is added to the report (this is one level higher than fnRptSectionHeading)
}
\description{
Add a section heading (level 3) to a report

Add a section title (level 2) to a report
}
\examples{
fnRptSectionHeading(rptVar = rpt$All, headingTitle = "Grade Distributions")
fnRptSectionHeading(rptVar = rpt$All, headingTitle = "Grade Distributions")
}
