% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTagAnalysis.R
\name{fnTagAnalysis}
\alias{fnTagAnalysis}
\title{Tag Analysis Table and Plot}
\usage{
fnTagAnalysis(
  variableName = NULL,
  data = NULL,
  variableOptions = NULL,
  variableNQuestions = NULL,
  scoreOptions = NULL,
  textLabels = NULL
)
}
\arguments{
\item{variableName}{***}

\item{data}{***}

\item{variableOptions}{***}

\item{variableNQuestions}{The number of times a question on each variable occurs. The first col contains the variables (should be titled "Var1") and the second col contains the number of times each occurs (should be titled "Number\nof\nQuestions")}

\item{scoreOptions}{***}

\item{textLabels}{Boolean (TRUE/FALSE). Set to TRUE to add the specific tag names to the x-axis on the plot. Otherwise each tag will be given a number (with a corresponding number also in the generated table)}
}
\value{
***
}
\description{
Tag Analysis Table and Plot
}
\examples{
***


}
