% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTagAnalysis.R
\name{fnTagAnalysis}
\alias{fnTagAnalysis}
\title{Tag Analysis Table and Plot}
\usage{
fnTagAnalysis(
  data = NULL,
  scoreOptions = NULL,
  variableName = NULL,
  variableOptions = NULL,
  variableNQuestions = NULL
)
}
\arguments{
\item{data}{***}

\item{scoreOptions}{***}

\item{variableName}{***}

\item{variableOptions}{***}

\item{variableNQuestions}{The number of times a question on each variable occurs. The first col contains the variables (should be titled "Var1") and the second col contains the number of times each occurs (should be titled "Number\nof\nQuestions")}
}
\value{
***
}
\description{
Tag Analysis Table and Plot
}
\examples{
***


}
