% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTagAnalysis.R
\name{fnTagAnalysis}
\alias{fnTagAnalysis}
\title{Tag Analysis Table and Plot}
\usage{
fnTagAnalysis(
  data = NULL,
  variableName = NULL,
  variableOptions = NULL,
  variableNQuestions = NULL,
  scoreOptions = NULL,
  textLabels = NULL,
  savePlot = NULL,
  plotsFolder = NULL
)
}
\arguments{
\item{data}{***}

\item{variableName}{***}

\item{variableOptions}{***}

\item{variableNQuestions}{The number of times a question on each variable occurs. The first col contains the variables (should be titled "Var1") and the second col contains the number of times each occurs (should be titled "Number\nof\nQuestions")}

\item{scoreOptions}{***}

\item{textLabels}{Boolean (TRUE/FALSE). Set to TRUE to add the specific tag names to the x-axis on the plot. Otherwise each tag will be given a number (with a corresponding number also in the generated table)}

\item{savePlot}{(Optional - if not set, plots are saved by default) A boolean (TRUE/FALSE, default TRUE) determines whether the plot will be saved. Useful for controlling overall save behaviour.}

\item{plotsFolder}{(Optional - if not set, no plots are saved but they are still created and returned) The root folder where images should be saved. Individual saved images are given unique names within that folder as part of this function}
}
\value{
A list of plots is returned. The list contains all the plots generated. To use the generated plots in your working environment, follow the example which uses the append function. Separately, if requested, the generated plots are also saved to a specified folder.
}
\description{
Tag Analysis Table and Plot
}
\examples{
tagResults <- list() # Do not include if the list already exists, this will clear it
tagResults <- append(
tagResults,
fnTagAnalysis(
  data = tag$scoresAllNegSpeciality,
  variableName = "Speciality",
  variableOptions = tag$lSpecialities,
  variableNQuestions = tag$nQuestionsSpeciality,
  scoreOptions = tag$scoreOptions,
  textLabels = FALSE
  )
)


}
