% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnTestRetest.R
\name{fnTestRetest}
\alias{fnTestRetest}
\title{Test Retest Dataframe Creation}
\usage{
fnTestRetest(listOfDetails = NULL, results = NULL, resultsPrevious = NULL)
}
\arguments{
\item{listOfDetails}{This variable should be a list containing as least the following defined variables: assessment (the current assessment number - eg: PT36) and assessmentPrev (the previous assessment number - eg: PT35)}

\item{results}{A dataframe - containing results data for the current assessment with at least the following columns: "StudentID", "Stage", assessement score (of the form "assessement_Score" where assessment matches the variable defined in listOfDetails) and assessment grade (of the form "assessment_Grade" where assessment grade matched the variable defined in listOfDetails)}

\item{resultsPrevious}{A dataframe - containing results data for the previous assessment, in a similar fashion to the "results" dataframe}
}
\value{
A dataframe is returned. It contains the student ID, stage and current and previous assessment grades and scores.
}
\description{
Test Retest Dataframe Creation
}
\examples{
dfTestRetest <- fnTestRetest(listOfDetails = cnst, results = dfRes, resultsPrevious = dfPrevTestData)
}
